requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
feedparser
beautifulsoup4
requests
apscheduler
python-dotenv
pydantic
google-genai
deep-translator
